+++ AppTinyMesh/Include/meshcolor.h
@@ -8,17 +8,17 @@ class MeshColor : public Mesh
 {
 protected:
   std::vector<Color> colors; //!< Array of colors.
-  std::vector<int> carray;  //!< Indexes.
+  std::vector<size_t> carray;  //!< Indexes.
 
 public:
   explicit MeshColor();
   explicit MeshColor(const Mesh&);
-  explicit MeshColor(const Mesh&, const std::vector<Color>&, const std::vector<int>&);
+  explicit MeshColor(const Mesh&, const std::vector<Color>&, const std::vector<size_t>&);
   ~MeshColor();
 
   Color GetColor(int) const;
   std::vector<Color> GetColors() const;
-  std::vector<int> ColorIndexes() const;
+  std::vector<size_t> ColorIndexes() const;
 };
 
 /*!
@@ -42,7 +42,7 @@ inline std::vector<Color> MeshColor::GetColors() const
 /*!
 \brief Return the set of color indices.
 */
-inline std::vector<int> MeshColor::ColorIndexes() const
+inline std::vector<size_t> MeshColor::ColorIndexes() const
 {
   return carray;
 }
